move genesis tests, structure properly, add yield keys for future test-vec generator
